requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.2.3
pymongo==4.10.1
streamlit==1.40.1
